requirements.txt
cython
faiss-cpu>=1.6.1
filelock
numpy
regex
torch>=1.5.0
transformers==4.6.1
tqdm>=4.27
wget
jsonlines
pandas
nltk
attrs
tensorboard
pyserini
apex
ujson
gdown